Move conversion logic into ColorSpace (#8)
Add a `convert` method to `ColorSpace`, and migrate logic in the
individual color types to use that.

One consequence is that gradient interpolation in Oklch is expected to
be faster, as the conversion to Oklab for error evaluation now doesn't
go all the way back to linear sRGB.

Also adds examples from koloro crate, which were inadvertently omitted
from the initial commit.

Removes `TaggedColor::from_linear_srgb` method, as that functionality
can now be expressed more idiomatically with properly typed colors.