Add Adobe RGB (1998) color space (#35)
The Linear A98 RGB <-> Linear sRGB matrices are based on the exact
rational conversion matrices calculated from the XYZ chromaticities.
This crate doesn't need the exact rationals as they're truncated to f32
at compile-time anyway, but it might be worth having to serve as a
reference for other people. It would be nice to have agreement across
implementations at any given precision.

The same _can_ be done for other basis change matrices, though it
becomes somewhat unwieldy as the number of steps combined into the
matrix increases; by my calculation element (0,0) of the Linear sRGB to
XYZ-D50 matrix with linear Bradford adaptation is exactly
`7206783369626736730083375524119595656 /
16526827482518743502029642682152162875`.

This is all based on the assumption that the color spaces' specified xy
chromaticities are normative, and the specified matrices in the specs
are computed off those.