Improve interop with `bytemuck` and FFI. (#43)
`AlphaColor`, `OpaqueColor`, and `PremulColor` are now using
`#[repr(transparent)]` and impl the `bytemuck` traits `Pod`,
`TransparentWrapper`, and `Zeroable`.

`Rgba8` is now `#[repr(C)]` and impls the `bytemuck` traits `Pod` and
`Zeroable`.

The `bytemuck` traits are only available when the new `bytemuck` feature
is enabled. It is not enabled by default.

This requires relaxing the forbidding of `unsafe_code` as the impls of
the `bytemuck` traits must use `unsafe impl`.