Add ACEScg color space (#54)
This pull request resolves #14. See that issue topic and [this Zulip
topic](https://xi.zulipchat.com/#narrow/channel/466849-color/topic/White.20Point.20and.20Color.20Transforms)
for more information, especially on how that matrix was calculated.

I believe this is the first color space added to this crate that is not
part of the CSS 4 spec. Because of this, I'm not sure how parsing or
serialization should be done for this color space. Guidance would be
appreciated, though we could possibly address this in a future PR.