Add a color parse function consuming a string prefix (#69)
The unparsed remainder of the string is indicated in the return value.

Resolves #60.

The existing function `parse_color` now returns
`ParseError::ExpectedEndOfString` if there are trailing characters. Some
choices could be made differently:

- For `#f00fa` the error `ParseError::WrongNumberOfHexDigits` is
returned. It could instead parse as `#f00f` with trailing `a`. The
current implementation is consistent with parsing `ident` to the longest
string and then checking for matches.
- This drops `ParseError::InvalidHexDigit`, as it is ambiguous whether
the character was intended as a hex digit or is the start of the
trailing string. `#fffffftr` is parsed as `#ffffff` with trailing `tr`.
This causes a slightly worse error for the full-string-consuming
`parse_color`, namely `ParseError::ExpectedEndOfString`. We could have
`parse_color` still return `ParseError::InvalidHexDigit` for that case,
but then the two functions have slightly different parsing behavior,
which is not ideal (the prefix parse might succeed, whereas the full
parse fails with an error different than `ExpectedEndOfString`).