Add `bytemuck` support for `HueDirection`, and `ColorSpaceTag` (#81)
These now impl `CheckedBitPattern`, `Contiguous`, `NoUninit`, and
`Zeroable`. They do not impl `Pod` as they are only valid for a subset
of the bit patterns.

Additionally, these types are now `repr(u8)` and have directly assigned
integer values so that we can be sure that the impl of the `Contiguous`
trait is correct.

This will likely be useful when working with gradients in Vello and it
has to encode things to the draw stream.

It may also be useful when we're doing a C FFI in the future.