Serialize color names by packing them into `DynamicColor` (#39)
To perform correct CSS Color 4 serialization of colors, some details of
how a color was declared have to be preserved. Specifically, [CSS Color
Module Level 4 §
15.2](https://www.w3.org/TR/css-color-4/#serializing-sRGB-values) and
the two sections on Lab and Oklab following it specify that
- a color declared as a named CSS/X11 color, should serialize to that
(normalized) color name; and
- a color declared as one of the named color space functions – i.e.,
_not_ the `color()` function – should serialize to that function.
(Exceptions being `hsl()` and `hwb()`, those serialize to `rgb()`.)

This implementation uses the padding currently available on
`DynamicColor` to pack in the color name the `DynamicColor` was created
from, or whether a named color space function was used; an
implementation of the ideas in
https://github.com/linebender/color/issues/26#issuecomment-2467997421.

`DynamicColor` now contains `Flags`, and the existing `Missing` color
components are moved into `Flags`. In this implementation, only
`parse_color` sets the color name or color space function name.

---------

Co-authored-by: Daniel McNab <36049421+DJMcNab@users.noreply.github.com>