Move docs on Main Types to the top (#97)
Moves the "Main types" section above the "Scope and goals" section in
the lib.rs docs.

Justification: I think the documentation on "what are the main types in
this crate" / "how do I use this crate" ought to be front and center.
Whereas the "background" on scope and goal is more secondary. The "Main
Types" section is also quite a bit shorter than the "Scope and goals"
section, so it means both fit "above the fold".

More concretely, I was reading the docs and trying to work out which
type(s) could be considered the primary/default color representation(s).
And it took me quite a bit of time to find it even though there was
already a section explaining it.

---------

Signed-off-by: Nico Burns <nico@nicoburns.com>