Don't enable serde's `std` feature (#108)
When we use the `serde?/std` syntax for the feature, we end up with a
public dependency on `serde` even when it isn't enabled or used. It
isn't compiled, but it feels wrong. We don't need the `std` feature from
`serde`, so no need to enable it and someone that does will already have
it enabled.